Removed warnings in MemoryManager and PoolAllocator

+++ MemoryManager/MemoryManager.cpp
@@ -14,14 +14,14 @@ MemoryManager::~MemoryManager()
 
 }
 
-PoolAllocator* MemoryManager::addPool(size_t sizeBytesEachEntry, unsigned int numEntries)
+PoolAllocator* MemoryManager::addPool(unsigned int sizeBytesEachEntry, unsigned int numEntries)
 {
 	PoolAllocator* temp = new PoolAllocator(getMem(sizeBytesEachEntry * numEntries), sizeBytesEachEntry, numEntries);
 	m_pools.push_back(temp);
 	return m_pools.back();
 }
 
-StackAllocator* MemoryManager::addStack(size_t sizeBytes)
+StackAllocator* MemoryManager::addStack(unsigned int sizeBytes)
 {
 	m_stacks.push_back(new StackAllocator(getMem(sizeBytes), sizeBytes));
 	return m_stacks.back();
